requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-RESTX==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# API Documentation
flask-restx==1.1.0
aniso8601>=0.82
jsonschema>=3.0.1
pytz>=2017.2

# Monitoring
prometheus-client==0.19.0
prometheus-flask-exporter==0.22.4
redis==4.6.0

# Database
SQLAlchemy==2.0.21
alembic==1.12.1
psycopg2-binary==2.9.9  # For PostgreSQL

# Web Scraping & APIs
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
praw==7.7.1

# Data Processing & ML
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.2
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3.post1

# Development
pytest==7.4.2
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
gevent==23.7.0
playwright
undetected-chromedriver
scikit-learn
praw
python-dotenv
clipboard
PyYAML>=6.0.1
